refactored Task model and added POST endpoint for creating tasks
