+++ src/db/mongoose.js
@@ -1,20 +1,8 @@
+// Connects to the database
+
 const mongoose = require('mongoose');
 
 mongoose.connect('mongodb://127.0.0.1:27017/task-manager-api', {
     useNewUrlParser: true, 
     useCreateIndex: true
-})
-
-// Define tasks model
-// pass in name and then define the model
-const Task = mongoose.model('Task', {
-    description: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    completed: {
-        type: Boolean,
-        default: false
-    }
 })
