+++ src/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 require('./db/mongoose');
 const User = require('./models/user');
+const Task = require('./models/task');
 
 const app = express();
 const port = process.env.PORT || 3000;
@@ -15,7 +16,17 @@ app.post('/users', (req, res) => {
     const user = new User(req.body);
 
     user.save().then(() => {
-        res.send(user);
+        res.status(201).send(user);
+    }).catch((error) => {
+        res.status(400).send(error);
+    })
+})
+
+app.post('/tasks', (req, res) => {
+    const task = new Task(req.body);
+    
+    task.save().then (() => {
+        res.status(201).send(task)
     }).catch((error) => {
         res.status(400).send(error);
     })
