+++ src/models/task.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose');
+
+// Define tasks model
+// pass in name and then define the model
+const Task = mongoose.model('Task', {
+    description: {
+        type: String,
+        required: true,
+        trim: true
+    },
+    completed: {
+        type: Boolean,
+        default: false
+    }
+})
+
+module.exports = Task;
